use JSON file, use vector for point display
discarded API call solutions due to AJAX limitations and the difficulty
to migrate this project out of a server. (So this app is just a mockup
using a sample xml). Could consider update xml periodically, or directly
make PHP script call Google API, but this will not happen on RIT's
Gibson.
Due to the limitation of can't display multiple markers on OSM, using a
work around to make markers as vectors. This is a great limitation of
OSM.